requirements-snowflake-connector.txt
asn1crypto==1.5.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==39.0.0
DateTime==5.2
filelock==3.9.0
greenlet==3.0.1
idna==3.4
numpy==1.26.1
oscrypto==1.3.0
pandas==2.1.2
pycparser==2.21
pycryptodomex==3.16.0
PyJWT==2.6.0
pyOpenSSL==23.0.0
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
six==1.16.0
snowflake-connector-python==3.0.0
snowflake-sqlalchemy==1.5.0
SQLAlchemy==1.4.50
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.14
zope.interface==6.1